Fix 'tester plugin is not active' errors
